tutorials/android: Update for NDK r18 changes
libc++_shared needs to be enabled in the application in order for it to
be included and used correctly for c++ plugins.

Only symbols from libraries explicitly mentioned are exported from
libgstreamer_android.so so we need to add a couple of deps to be able to
access the necessary functions.

Requires
https://gitlab.freedesktop.org/gstreamer/cerbero/merge_requests/31
https://gitlab.freedesktop.org/gstreamer/cerbero/merge_requests/1