ci_template: blacklist a problematic test in check
Also pass --dump-on-failure in the job. While the logs are
artifacts and they are integrated in the MRs, its not as
convinient to findout out what is failling if you are looking
into just a job log, of a commit. Thus make it show logs are also
priting on stdout/stderr.