meson: detect opengl api from -base .pc files correctly
There was a mismatch between the .pc files generated by
autotools and by meson that would lead to meson not detecting
that opengl api is available even though it is, if -base was
built with autotools. The mismatch has now been rectified in
-base, so we need to update for that.

This is mostly for consistency, this problem didn't seem
to affect anything in -good.

See https://gitlab.freedesktop.org/gstreamer/gst-plugins-bad/issues/871