gst-inspect: Re-add DEFAULT_LESS_OPTS with initial value
Commit 56b4fbef5e6760adc927d0e1c7c8d6a0db9b785c refactored the pipe code
to use GLib utility, but the patch was hading some other changed. LESS
env was now hardcoded in the middle instead of from a define and was
changed from FXR to -RX. The "-" is not even valid for LESS env, and
with the lost of F, we would still use a pager when the content fits the
terminal.